+++ frontend/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feature: multi select for organisations
 - bugfix: allow multiple query parameters to be set
+- bugfix: select check all methods could return 0 instead of a boolean
 - chore: bump @angular-eslint/eslint-plugin to 18.3.1
 - chore: bump typescript-eslint to 8.6.0
 - chore: bump @typescript/eslint-plugin to 8.14.0